pyproject.toml
[tool.poetry]
name = "filter_sam"
version = "0.2.0"
description = "filter lines in a sam file that do not have a matching entry in a fasta file"
authors = ["Miles Smith <mileschristiansmith@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.rst"
classifiers = [
        "Development Status :: 2 - Pre-Alpha",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: BSD License",
        "Natural Language :: English",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
]
repository = "https://github.com/milescsmith/filter_sam"
include = [ "HISTORY.rst" ]

[tool.poetry.scripts]
"filter_sam" = "filter_sam.cli:app"

[tool.poetry.dependencies]
python = ">=3.8"
click = "^8.0.3"
biopython = "^1.79"
pysam = "^0.18.0"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
mypy = "^0.930"
flake8 = "nox"
nox = "^2021.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
